Start setting up resource locks

+++ src/migrations/2025-04-22T00:15:56+00:00.ts
@@ -4,7 +4,7 @@ export async function up(db: Kysely<any>): Promise<void> {
     await db.schema
         .createTable('AlpacaAccount')
         .ifNotExists()
-        .addColumn('id', 'integer', (col) => col.primaryKey().autoIncrement())
+        .addColumn('id', 'bigint', (col) => col.primaryKey().autoIncrement())
         .addColumn('platformAccountId', 'text', (col) => col.notNull())
         .addColumn('platformAPIKey', 'text', (col) => col.notNull())
         .execute();
